# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: LeaseMind App Foundation API
  version: 1.2.0
  description: >-
    Synthetic-only contract for the seven HTTP operations that actually
    exist in apps/api/src/app.ts: two health probes, the Campaign Current
    State Projection read model, the Technical Assignment save-draft and
    read commands, and the atomic Campaign launch command (see
    03_ARCHITECTURE/decisions/ADR-0001-application-development-stack.md,
    ADR-0002-campaign-event-stream.md, ADR-0003-pre-launch-runtime-safety-gate.md,
    ADR-0004-read-only-api-contract.md,
    ADR-0007-synthetic-campaign-creation-command-boundary.md and
    ADR-0008-technical-assignment-implementation.md). No other write
    operation exists or is documented. PRODUCTION_LAUNCH_GATE remains
    blocked; this document does not authorize any production deployment,
    real personal data, real payments or protected-data disclosure.
servers:
  - url: http://127.0.0.1:3001
    description: >-
      Synthetic local development server only (see apps/api/.env.example).
      No production server exists.
paths:
...
  /api/v1/campaigns:
    get:
      operationId: listCampaigns
      summary: List all synthetic Campaigns
      description: >-
        Returns every row of the derived, rebuildable Campaign Current
        State Projection (apps/api/migrations/001_campaign_current_state_projection.up.sql),
        ordered by campaign_id ascending. Read-only; see the POST operation
        below for the one write operation on this path.
      responses:
        '200':
          description: The full list of synthetic Campaigns.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignListResponse'
        '503':
          description: The database is unreachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUnavailableResponse'

    post:
      operationId: launchCampaign
      summary: Atomically launch a new synthetic Campaign from a ready Technical Assignment
      description: >-
        The only write operation on this path. One atomic transaction:
        verifies the referenced Technical Assignment is
        `lifecycle_status=ready_for_analysis` at exactly `expected_revision`
        and that a valid Contacts Gate marker is present, then creates the
        synthetic Campaign in status `Created`, appends the
        `campaign.subject_linked.v1` provenance event, and marks the
        Technical Assignment `campaign_started`
        (apps/api/src/db/launchCampaign.ts). campaign_id, aggregate_version,
        created_at and updated_at are always server-generated; campaign_id
        is derived deterministically from idempotency_key, so retrying the
        same command with the same idempotency_key returns the same
        Campaign instead of creating a duplicate (200, not 201). The
        request body accepts no name, phone, email or any other contact
        field. See ADR-0007-synthetic-campaign-creation-command-boundary.md
        and ADR-0008-technical-assignment-implementation.md.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LaunchCampaignRequest'
      responses:
        '201':
          description: A new synthetic Campaign was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '200':
          description: >-
            idempotency_key was already used for this exact command; the
            existing Campaign is returned unchanged, no duplicate created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '400':
          description: >-
            The request body is malformed, the Technical Assignment is not
            ready/found, its revision does not match expected_revision, or
            the Contacts Gate marker is missing/invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LaunchCampaignErrorResponse'
        '503':
          description: The database is unreachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUnavailableResponse'
  /api/v1/technical-assignments:
    post:
      operationId: saveTechnicalAssignmentDraft
      summary: Create or update a synthetic Technical Assignment draft
      description: >-
        Creates a new draft on first call for a given idempotency_key, or
        updates the same draft on subsequent calls with the same key.
        Exactly 30 approved Property fields (`scenario=need_tenant`) or 29
        approved TenantRequest fields (`scenario=need_property`) are
        accepted -- no other property, including any name/phone/email/
        contact field, can be submitted. `technical_assignment_id`,
        `revision` and `lifecycle_status` are always server-computed.
        `property_exact_address` is accepted but never returned in any
        response (see `has_exact_address`). See
        02_PRODUCT/CAMPAIGN_TECHNICAL_ASSIGNMENT.md and
        ADR-0008-technical-assignment-implementation.md.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveTechnicalAssignmentRequest'
      responses:
        '201':
          description: A new Technical Assignment draft was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechnicalAssignment'
        '200':
          description: An existing Technical Assignment draft was updated (or an identical resubmission was a no-op).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechnicalAssignment'
        '400':
          description: The request body is malformed, a field is invalid, a cross-field rule is violated, or the scenario does not match a prior save under this idempotency_key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechnicalAssignmentErrorResponse'
        '503':
          description: The database is unreachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUnavailableResponse'
  /api/v1/technical-assignments/{technicalAssignmentId}:
    get:
      operationId: getTechnicalAssignmentById
      summary: Get a single Technical Assignment draft by id
      parameters:
        - name: technicalAssignmentId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UuidV4OrV7'
      responses:
        '200':
          description: The matching Technical Assignment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechnicalAssignment'
        '404':
          description: No Technical Assignment exists with the given id.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [error]
                properties:
                  error:
                    type: string
                    enum: [TECHNICAL_ASSIGNMENT_NOT_FOUND]
  /api/v1/campaigns/{campaignId}:
    get:
      operationId: getCampaignById
      summary: Get a single synthetic Campaign by id
      description: >-
        campaignId must be a UUID v4 or v7; any other syntactically valid
        UUID version (v1, v2, v3, v5, v6, v8) or malformed value is
        rejected with 400 before any database query runs
        (apps/api/src/uuid.ts).
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UuidV4OrV7'
      responses:
        '200':
          description: The matching Campaign.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '400':
          description: campaignId is not a well-formed UUID v4 or v7.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidCampaignIdResponse'
        '404':
          description: No Campaign exists with the given campaignId.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignNotFoundResponse'
        '503':
          description: The database is unreachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUnavailableResponse'
components:
  schemas:
...
    ServiceUnavailableResponse:
      type: object
      additionalProperties: false
      required: [status, database]
      properties:
        status:
          type: string
          enum: [error]
        database:
          type: string
          enum: [unreachable]
    Campaign:
      type: object
      additionalProperties: false
      required: [campaign_id, status, aggregate_version, created_at, updated_at]
      properties:
        campaign_id:
          type: string
          format: uuid
        status:
          type: string
          description: >-
            One of the 11 approved Campaign statuses
            (LeaseMind_AI_MANAGER_ARCHITECTURE_v1.0.md, Approved v1.0,
            section 5.3). No additions, renames or removals.
          enum:
            - Created
            - Analyzing
            - Strategy Building
            - Searching
            - Qualifying
            - Negotiating
            - Viewing
            - Deal Support
            - Completed
            - Paused
            - Failed
        aggregate_version:
          type: string
          pattern: '^[1-9][0-9]*$'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CampaignListResponse:
      type: object
      additionalProperties: false
      required: [campaigns]
      properties:
        campaigns:
          type: array
          items:
            $ref: '#/components/schemas/Campaign'
...
    LaunchCampaignRequest:
      type: object
      additionalProperties: false
      required: [idempotency_key, technical_assignment_id, expected_revision, contacts_gate_evidence]
      properties:
        idempotency_key:
          type: string
          minLength: 1
          maxLength: 200
          description: >-
            Caller-generated deduplication token for this launch command.
            Never a contact field -- no other property is accepted on this
            request.
        technical_assignment_id:
          $ref: '#/components/schemas/UuidV4OrV7'
        expected_revision:
          type: integer
          minimum: 1
          description: The Technical Assignment revision the caller last observed as ready_for_analysis; a mismatch means the Analysis the caller saw is stale.
        contacts_gate_evidence:
          type: string
          minLength: 1
          maxLength: 100
          description: Synthetic-only Contacts Gate marker (ADR-0008 section 2). A real consent record is a Launch blocker, not implemented.
    LaunchCampaignErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
          enum:
            - INVALID_IDEMPOTENCY_KEY
            - TECHNICAL_ASSIGNMENT_STATE_INVALID
            - TECHNICAL_ASSIGNMENT_ANALYSIS_REQUIRED
            - TECHNICAL_ASSIGNMENT_ANALYSIS_STALE
            - TECHNICAL_ASSIGNMENT_CONTACTS_GATE_REQUIRED
            - TECHNICAL_ASSIGNMENT_REVISION_CONFLICT
    TechnicalAssignmentErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        field_id:
          type: [string, 'null']
        error:
          type: string
          enum:
            - TECHNICAL_ASSIGNMENT_SCHEMA_UNSUPPORTED
            - TECHNICAL_ASSIGNMENT_SCENARIO_REQUIRED
            - TECHNICAL_ASSIGNMENT_SCENARIO_IMMUTABLE
            - TECHNICAL_ASSIGNMENT_STATE_INVALID
            - TECHNICAL_ASSIGNMENT_REQUIRED_FIELD_MISSING
            - TECHNICAL_ASSIGNMENT_FIELD_TYPE_INVALID
            - TECHNICAL_ASSIGNMENT_FIELD_VALUE_INVALID
            - TECHNICAL_ASSIGNMENT_CROSS_FIELD_CONFLICT
            - TECHNICAL_ASSIGNMENT_PERSONAL_DATA_FORBIDDEN
            - TECHNICAL_ASSIGNMENT_SECRET_FORBIDDEN
    SaveTechnicalAssignmentRequest:
      type: object
      additionalProperties: false
      required: [idempotency_key, scenario, payload]
      properties:
        idempotency_key:
          type: string
          minLength: 1
          maxLength: 200
        scenario:
          type: string
          enum: [need_tenant, need_property]
        payload:
          description: Exactly the 30 Property fields (need_tenant) or 29 TenantRequest fields (need_property) -- see PropertyPayload / TenantRequestPayload. No other property is accepted.
          oneOf:
            - $ref: '#/components/schemas/PropertyPayload'
            - $ref: '#/components/schemas/TenantRequestPayload'
    TechnicalAssignment:
      type: object
      additionalProperties: false
      required: [technical_assignment_id, scenario, schema_version, revision, lifecycle_status, has_exact_address, payload, created_at, updated_at]
      properties:
        technical_assignment_id:
          type: string
          format: uuid
        scenario:
          type: string
          enum: [need_tenant, need_property]
        schema_version:
          type: string
          enum: ['1.0']
        revision:
          type: integer
          minimum: 1
        lifecycle_status:
          type: string
          enum: [draft, ready_for_analysis, campaign_started]
        has_exact_address:
          type: boolean
          description: True if a protected property_exact_address is stored. The raw address itself is never included in any response.
        payload:
          oneOf:
            - $ref: '#/components/schemas/PropertyPayload'
            - $ref: '#/components/schemas/TenantRequestPayload'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...